refactor: 영수증 위젯 안에 작성된 세부 위젯들 별도 파일로 분리해서 가독성 개선
