+++ lib/widgets/receipt.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:daily_receipt/models/todos.dart';
+import 'receipt_edge_clipper.dart';
+import 'receipt_item.dart';
+import 'receipt_text.dart';
 
 class ReceiptComponent extends StatelessWidget {
   final List<Todo> todos;
@@ -89,71 +92,4 @@ class ReceiptComponent extends StatelessWidget {
       ),
     );
   }
-}
-
-class ReceiptItem extends StatelessWidget {
-  final String name;
-  final String value;
-
-  const ReceiptItem(this.name, this.value, {super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        ReceiptText(name),
-        ReceiptText(value),
-      ],
-    );
-  }
-}
-
-class ReceiptText extends StatelessWidget {
-  final String text;
-
-  const ReceiptText(this.text, {super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      text,
-      style: Theme.of(context).textTheme.titleSmall?.copyWith(
-            color: Theme.of(context).colorScheme.primary,
-          ),
-    );
-  }
-}
-
-class ReceiptClipper extends CustomClipper<Path> {
-  @override
-  Path getClip(Size size) {
-    final path = Path();
-    const wavelength = 14.0;
-    const amplitude = 5.0;
-
-    // 상단 톱니
-    path.moveTo(0, amplitude);
-    for (var i = 0; i < size.width / wavelength; i++) {
-      path.lineTo((i + 0.5) * wavelength, 0);
-      path.lineTo((i + 1) * wavelength, amplitude);
-    }
-
-    // 오른쪽 변
-    path.lineTo(size.width, size.height - amplitude);
-
-    // 하단 톱니
-    for (var i = size.width / wavelength; i > 0; i--) {
-      path.lineTo((i - 0.5) * wavelength, size.height);
-      path.lineTo((i - 1) * wavelength, size.height - amplitude);
-    }
-
-    // 왼쪽 변
-    path.lineTo(0, amplitude);
-
-    return path;
-  }
-
-  @override
-  bool shouldReclip(CustomClipper<Path> oldClipper) => false;
 }
